Probbebly so data type/endian issues going on

+++ includes/types.hpp
@@ -9,7 +9,7 @@
 # include <stdint.h>
 
 class							Debugee;
-typedef size_t					t_word;
-typedef size_t					t_program_ptr;
+typedef int64_t					t_word;
+typedef int64_t					t_program_ptr;
 typedef uint64_t				t_reg;
 #endif // TYPES_HPP
